add: 캠핑 유형 컴포넌트 분리
