+++ src/app/(pages)/meets/components/meetsWrite/MeetCreatorForm.tsx
@@ -5,7 +5,8 @@ import DropDownCampSearch from "./DropDownCampSearch";
 import CDateRangePicker from "@/_components/date/CDateRangePicker";
 import { MeetWithCamp } from "../../types/meet.types";
 import useMeetCreatorForm from "../../hooks/useMeetCreatorForm";
-import MeetCreatorInfoSection from "./MeetCreatorInfoSection";
+import MeetCreatorInfoSection from "./meetCreatorSection/MeetCreatorInfoSection";
+import MeetCreatorTypeSection from "./meetCreatorSection/MeetCreatorTypeSection";
 
 const MeetCreatorForm = ({ meetWithCamp }: { meetWithCamp: MeetWithCamp }) => {
   const {
@@ -25,13 +26,6 @@ const MeetCreatorForm = ({ meetWithCamp }: { meetWithCamp: MeetWithCamp }) => {
     isOpen,
     searchList
   } = useMeetCreatorForm(meetWithCamp);
-
-  const groundTypes = {
-    crushedStone: "파쇄석",
-    woodenDeck: "나무데크",
-    soil: "흙",
-    any: "상관없음"
-  };
 
   const showDropDown = isOpen && !!searchList && searchList.length > 0;
 
@@ -55,7 +49,9 @@ const MeetCreatorForm = ({ meetWithCamp }: { meetWithCamp: MeetWithCamp }) => {
 
         <MeetCreatorInfoSection />
 
-        <section className="mt-[123px]">
+        <MeetCreatorTypeSection watch={watch} setValue={setValue} />
+
+        {/* <section className="mt-[123px]">
           <h2 className="mb-[70px] text-xl font-[500]">캠핑 유형</h2>
           <div className="flex gap-[14px] border-b-[1px] border-[#D4D4D4] pb-[37px] font-[15px]">
             {Object.entries(groundTypes).map(([key, value]) => (
@@ -66,8 +62,7 @@ const MeetCreatorForm = ({ meetWithCamp }: { meetWithCamp: MeetWithCamp }) => {
                   // 선택된 값으로 설정
                   setValue("ground_type", key);
                 }}
-                className={`h-[45px] rounded-[22.5px] border-[1px] border-[#C3C3C3] px-4 py-[14px] pb-2 pl-7 pr-7 pt-2 text-[#A4A4A4] ${watch("ground_type") === key ? "bg-[#D9D9D9] font-[600]" : ""}`}
-              >
+                className={`h-[45px] rounded-[22.5px] border-[1px] border-[#C3C3C3] px-4 py-[14px] pb-2 pl-7 pr-7 pt-2 text-[#A4A4A4] ${watch("ground_type") === key ? "bg-[#D9D9D9] font-[600]" : ""}`}>
                 {value}
               </button>
             ))}
@@ -79,8 +74,7 @@ const MeetCreatorForm = ({ meetWithCamp }: { meetWithCamp: MeetWithCamp }) => {
                 // 선택된 값으로 설정
                 setValue("is_newbie", true);
               }}
-              className={`h-[45px] rounded-[22.5px] border-[1px] border-[#C3C3C3] px-4 py-[14px] pb-2 pl-7 pr-7 pt-2 text-[#A4A4A4] ${watch("is_newbie") === true ? "bg-[#D9D9D9] font-[600]" : ""}`}
-            >
+              className={`h-[45px] rounded-[22.5px] border-[1px] border-[#C3C3C3] px-4 py-[14px] pb-2 pl-7 pr-7 pt-2 text-[#A4A4A4] ${watch("is_newbie") === true ? "bg-[#D9D9D9] font-[600]" : ""}`}>
               초보가능
             </button>
             <button
@@ -89,12 +83,11 @@ const MeetCreatorForm = ({ meetWithCamp }: { meetWithCamp: MeetWithCamp }) => {
                 // 선택된 값으로 설정
                 setValue("is_newbie", false);
               }}
-              className={`h-[45px] rounded-[22.5px] border-[1px] border-[#C3C3C3] px-4 py-[14px] pb-2 pl-7 pr-7 pt-2 text-[#A4A4A4] ${watch("is_newbie") === false ? "bg-[#D9D9D9] font-[600]" : ""}`}
-            >
+              className={`h-[45px] rounded-[22.5px] border-[1px] border-[#C3C3C3] px-4 py-[14px] pb-2 pl-7 pr-7 pt-2 text-[#A4A4A4] ${watch("is_newbie") === false ? "bg-[#D9D9D9] font-[600]" : ""}`}>
               숙련자
             </button>
           </div>
-        </section>
+        </section> */}
 
         <div className="mb-20 w-[600px]">
           {errors.contentId && <span>캠핑장을 선택하세요.</span>}
