--- src/app/(pages)/meets/components/meetsWrite/MeetCreatorInfo.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const MeetCreatorInfo = () => {
-  return <div>MeetCreatorInfo</div>;
-};
-
-export default MeetCreatorInfo;
